Refactors the video player's card stack architecture to decouple generic swipe logic from reel-specific video playback.
*   **refactor: Introduces a generic `SwipeableCardStack` component**
    *   Extracts core gesture handling, animation management, and stack layout into a reusable `SwipeableCardStack` and `SwipeableCardStackItem`.
    *   Introduces `SwipeableCardStackItemScope` to provide swipe state (direction, progress) to card content.
    *   Moves card transformation logic (scale, offset, rotation, shadow) into the generic item wrapper.

*   **feat: Implements `ReelSwipeableCardStack` for video-specific functionality**
    *   Specializes the generic stack to handle video playback synchronization using `PlaybackCoordinator`.
    *   Connects swipe gestures to video playback events (e.g., setting active index on swipe).
    *   Integrates `SwipeButtons` and specialized feedback overlays (e.g., "MAST" and "FLOP" indicators) for the reel experience.

*   **refactor: Updates `TournamentGameScaffoldScreen`**
    *   Removes `PrefetchVideoListenerImpl` and updates `FeedScreen` parameters to align with the refactored player.
    *   Simplifies reel player configuration within the tournament game UI.

*   **chore: Clean up and optimization**
    *   Renames `CardStackItem` to `ReelCardContent` to reflect its role as the content within the new generic stack.
    *   Refactors `YRALReelPlayerCardStack` to utilize the new `ReelSwipeableCardStack`.
    *   Optimizes card rendering to use `Offscreen` compositing only where necessary for transformations.
